Disable extension if Storefront > 2.3.0
